externals/sdsvtd/requirements.txt
torch
mmcv-full